Major Updates. See commit details.
Updated ReadMe and docs.

Added ReadMe files to ActivityFeed and Push folders in samples folder.

Updated samples to use embedded frameworks for the UsergridSDK instead of cocoapods.

Updated ActivityFeed sample app to not include the Pods folder so we don't actually check in the 3rd party code used in the sample.

Added persistCurrentUserInKeychain to UsergridClientConfig.

Now saving the whole UsergridDevice object instead of just the UUID.

Added missing methods to UsergridClient, UsergridDevice and UsergridUser classes.