Updates. See commit details.
UsergridClient.getConnection methods now use a default value for queries in the method definition.

Removed invalid method UsergridDevice.uuid.

Added use of `guard` statement in favor of if/else in various parts of the code.

UsergridUser.create() now copies the properties from the response over to the User that called it.

UsergridUser.reauthenticate() now resets the UsergridClient.currentUser property if the calling User is in fact the same as the currentUser.

UsergridUser.connectToDevice() and UsergridUser.disconnectFromDevice() methods fixed.

UsergridUser.getConnectedDevice() methods added.

UsergridClient.logoutUser() now determines if the user to logout is in fact the currentUser and if so now sets that to nil if need be.

Decoding of UsergridClientConfig.authFallback is fixed.

Updates to tests.